MAPLEZ_LLM_report_labeler/classifier_src/requirements.txt
h5py==3.9.0
imageio==2.31.3
matplotlib==3.8.0
numpy==1.25.2
pandas==2.1.0
pillow==9.4.0
pytorch==2.0.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
torchvision==0.15.2
torchxrayvision==1.2.0
tqdm==4.66.1